requirements.txt
ipywidgets==7.7.1
matplotlib==3.7.0
numpy==1.24.2
pandas==1.5.3
tqdm==4.64.0
python-dotenv==1.0.0
openai>=0.28.0
tiktoken
google-api-python-client>=2.98.0,<3.0.0
youtube-transcript-api>=0.6.1,<1.0.0
rich>=13.5.2,<14.0.0